added http triggered function, refactored all graph/logging into a single class which is used by both the timer and http function
