+++ GraphTestFunction.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.9.34728.123
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "GraphTestFunction", "GraphTestFunction\GraphTestFunction.csproj", "{30287979-853A-4949-98D3-6988D359DA37}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "GraphTestFunction", "GraphTestFunction\GraphTestFunction.csproj", "{30287979-853A-4949-98D3-6988D359DA37}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
